+++ include/pinocchio/bindings/python/multibody/model.hpp
@@ -108,14 +108,22 @@ namespace pinocchio
             "rotorInertia", &Model::rotorInertia, "Vector of rotor inertia parameters.")
           .def_readwrite(
             "rotorGearRatio", &Model::rotorGearRatio, "Vector of rotor gear ratio parameters.")
-          .def_readwrite("friction", &Model::upperDryFrictionLimit, "Deprecated member. Vector of joint friction parameters.")
-          .def_readwrite("upperDryFrictionLimit", &Model::upperDryFrictionLimit, "Vector of maximum joint friction.")
-          .def_readwrite("lowerDryFrictionLimit", &Model::lowerDryFrictionLimit, "Vector of minimum joint friction.")
+          .def_readwrite(
+            "friction", &Model::upperDryFrictionLimit,
+            "Deprecated member. Vector of joint friction parameters.")
+          .def_readwrite(
+            "upperDryFrictionLimit", &Model::upperDryFrictionLimit,
+            "Vector of maximum joint friction.")
+          .def_readwrite(
+            "lowerDryFrictionLimit", &Model::lowerDryFrictionLimit,
+            "Vector of minimum joint friction.")
           .def_readwrite("damping", &Model::damping, "Vector of joint damping parameters.")
-          .def_readwrite("effortLimit", &Model::upperEffortLimit, "Deprecated member. Joint max effort.")
+          .def_readwrite(
+            "effortLimit", &Model::upperEffortLimit, "Deprecated member. Joint max effort.")
           .def_readwrite("upperEffortLimit", &Model::upperEffortLimit, "Joint max effort.")
           .def_readwrite("lowerEffortLimit", &Model::lowerEffortLimit, "Joint min effort.")
-          .def_readwrite("velocityLimit", &Model::upperVelocityLimit, "Deprecated member. Joint max velocity.")
+          .def_readwrite(
+            "velocityLimit", &Model::upperVelocityLimit, "Deprecated member. Joint max velocity.")
           .def_readwrite("lowerVelocityLimit", &Model::lowerVelocityLimit, "Joint min velocity.")
           .def_readwrite("upperVelocityLimit", &Model::upperVelocityLimit, "Joint max velocity.")
           .def_readwrite(
@@ -157,8 +165,9 @@ namespace pinocchio
           .def(
             "addJoint", &ModelPythonVisitor::addJoint2,
             bp::args(
-              "self", "parent_id", "joint_model", "joint_placement", "joint_name", "min_effort", "max_effort", "min_velocity",
-              "max_velocity", "min_config", "max_config", "min_friction", "max_friction", "damping"),
+              "self", "parent_id", "joint_model", "joint_placement", "joint_name", "min_effort",
+              "max_effort", "min_velocity", "max_velocity", "min_config", "max_config",
+              "min_friction", "max_friction", "damping"),
             "Adds a joint to the kinematic tree with given bounds. The joint is defined by its "
             "placement relative to its parent joint and its name.\n"
             "This signature also takes as input effort, velocity limits as well as the bounds on "
@@ -241,6 +250,8 @@ namespace pinocchio
 
           .PINOCCHIO_ADD_STATIC_PROPERTY_READONLY_BYVALUE(
             Model, gravity981, "Default gravity field value on the Earth.");
+
+        bp::register_ptr_to_python<std::shared_ptr<Model>>();
       }
 
       static JointIndex addJoint0(
@@ -286,8 +297,8 @@ namespace pinocchio
         const VectorXs & damping)
       {
         return model.addJoint(
-          parent_id, jmodel, joint_placement, joint_name, min_effort, max_effort, min_velocity, max_velocity, min_config,
-          max_config, min_friction, max_friction, damping);
+          parent_id, jmodel, joint_placement, joint_name, min_effort, max_effort, min_velocity,
+          max_velocity, min_config, max_config, min_friction, max_friction, damping);
       }
 
       ///